Pulses polarity and DAC improvements.
The pulses can be positive followed by a negative or the contrary.
A new bit is buffered with the value at the starting.
The state machine guarantee there is always, at least, one 0
  at the end of each pulse.
That is to implement the option of a totem-pole output,
  using 2 DACs per channel.
For more information, see the next commit.